Show draft docs that have been trashed but not yet committed in commit modal by default
This addresses https://github.com/apostrophecms/apostrophe/issues/1641.

When an editor trashes a document, that only affects the draft version
of that document. From a developer standpoint, this makes sense, since
the `trash` property is just another property on the doc in the database.
For the editor however, the document has effectively disappeared from the
site now, and it's common for them not to be aware of the fact that they
will need to find the document in the piece modal and manually commit it
from there for it to actually be removed from the live site.

This commit changes that so that apostrophe-workflow by default looks up
docs where the draft version has been trashed and the live version has
not when editors commit a document. This makes it easy and obvious that
trashed documents need to be committed to actually disappear.