Modify locale detection routine in the middleware
A new case, where a locale which shares hostname with other locales,
but doesn’t have a prefix (i.e. it exists at the root-level of the host), has
been handled in the locale detection routine. Note that when such a
locale doesn’t exist, the functionality remains unchanged/unaffected,
which makes this change fully backward compatible.