Fix a bug in cross-domain session implementation
When we serialize the session object and de-serialize when a user
visits a locale on another domain, the deserialized object becomes
a plain object. But express-session requires the session object to
be an instance of its Cookie prototype to function normally. This fix
ensures that the deserialized session object is re-instantiated as an
instance of the express-session Cookie prototype so that the sessions
are maintained properly across cross-domain locales.